+++ assets/radar-rules.js
@@ -1733,13 +1733,13 @@
         'trackings.post': [
             {
                 title: '郵便・荷物の追跡',
-                docs: 'https://docs.rsshub.app/other.html#ri-ben-you-bian',
+                docs: 'https://docs.rsshub.app/other.html#ri-ben-you-bian-you-bian-zhui-ji-サービス',
                 source: '/services/srv/search/direct',
                 target: (params, url) => {
                     const reqCode = new URL(url).searchParams.get('reqCodeNo1').toUpperCase();
                     const locale = new URL(url).searchParams.get('locale').toLowerCase();
                     if ((reqCode.search(/^(?:\d{12}|[A-Z]{2}\d{9}[A-Z]{2})$/) === 0 && locale === 'ja') || locale === 'en') {
-                        return `/japanpost/${reqCode}/${locale}`;
+                        return `/japanpost/track/${reqCode}/${locale}`;
                     }
                 },
             },
