+++ lib/router.js
@@ -2083,7 +2083,7 @@ router.get('/leboncoin/ad/:query', require('./routes/leboncoin/ad.js'));
 router.get('/dhl/:id', require('./routes/dhl/shipment-tracking'));
 
 // Japanpost
-router.get('/japanpost/:reqCode/:locale?', require('./routes/japanpost/index'));
+router.get('/japanpost/track/:reqCode/:locale?', require('./routes/japanpost/track'));
 
 // 中华人民共和国商务部
 router.get('/mofcom/article/:suffix', require('./routes/mofcom/article'));
